+++ coders/ps.c
@@ -860,7 +860,8 @@ static unsigned int WritePSImage(const ImageInfo *image_info,Image *image)
     text_size=0;
     attribute=GetImageAttribute(image,"Label");
     if (attribute != (ImageAttribute *) NULL)
-      text_size=MultilineCensus(attribute->value)*image_info->pointsize+12;
+      text_size=(unsigned int)
+	(MultilineCensus(attribute->value)*image_info->pointsize+12);
     width=image->columns;
     height=image->rows;
     x=0;
